add Numpy

+++ finchart/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+import numpy
 
 # Create your models here.
 class Company(models.Model):
@@ -36,57 +37,57 @@ class Fstatement(models.Model):
 
     # 流動資産比率（流動資産／総資産）
     def current_assets_rate(self):
-        f = self.bs_current_assets / self.bs_total_assets() * 100
+        f = numpy.float64(self.bs_current_assets) / self.bs_total_assets() * 100
         return f
 
     # 固定資産比率（固定資産／総資産）
     def fixed_assets_rate(self):
-        f = self.bs_fixed_assets / self.bs_total_assets() * 100
+        f = numpy.float64(self.bs_fixed_assets) / self.bs_total_assets() * 100
         return f
 
     # 繰延資産比率（繰延資産／総資産）
     def deferred_assets_rate(self):
-        f = self.bs_deferred_assets / self.bs_total_assets() * 100
+        f = numpy.float64(self.bs_deferred_assets) / self.bs_total_assets() * 100
         return f
 
     # 流動負債比率（流動負債／総資産）
     def current_liabilities_rate(self):
-        f = self.bs_current_liabilities / self.bs_total_assets() * 100
+        f = numpy.float64(self.bs_current_liabilities) / self.bs_total_assets() * 100
         return f
 
     # 固定負債比率（固定負債／総資産）
     def fixed_liabilities_rate(self):
-        f = self.bs_fixed_liabilities / self.bs_total_assets() * 100
+        f = numpy.float64(self.bs_fixed_liabilities) / self.bs_total_assets() * 100
         return f
 
     # 純資産比率（純資産／総資産）
     def net_assets_rate(self):
-        f = self.bs_net_assets / self.bs_total_assets() * 100
+        f = numpy.float64(self.bs_net_assets) / self.bs_total_assets() * 100
         return f
 
     # 売上総利益比率（売上総利益／総売上）
     def gross_profit_rate(self):
-        f = self.pl_gross_profit / self.pl_gross_sales * 100
+        f = numpy.float64(self.pl_gross_profit) / self.pl_gross_sales * 100
         return f
 
     # 営業利益比率（営業利益／総売上）
     def operating_profit_rate(self):
-        f = self.pl_operating_profit / self.pl_gross_sales * 100
+        f = numpy.float64(self.pl_operating_profit) / self.pl_gross_sales * 100
         return f
 
     # 経常利益比率（経常利益／総売上）
     def ordinary_income_rate(self):
-        f = self.pl_ordinary_income / self.pl_gross_sales * 100
+        f = numpy.float64(self.pl_ordinary_income) / self.pl_gross_sales * 100
         return f
 
     # 税引前当期純利益比率（税引前当期純利益／総売上）
     def income_before_tax_rate(self):
-        f = self.pl_income_before_tax / self.pl_gross_sales * 100
+        f = numpy.float64(self.pl_income_before_tax) / self.pl_gross_sales * 100
         return f
 
     # 当期純利益比率（当期純利益／総売上）
     def net_income_rate(self):
-        f = self.pl_net_income / self.pl_gross_sales * 100
+        f = numpy.float64(self.pl_net_income) / self.pl_gross_sales * 100
         return f
 
     # キャッシュフロー
@@ -97,17 +98,17 @@ class Fstatement(models.Model):
     # 流動比率を計算
     def current_rate(self):
         if self.bs_current_liabilities > 0:
-            f = self.bs_current_assets / self.bs_current_liabilities * 100
+            f = numpy.float64(self.bs_current_assets) / self.bs_current_liabilities * 100
         else:
             f = "-"
         return f
 
     # ＲＯＥを計算
     def roe(self):
-        f = self.pl_operating_profit / self.bs_net_assets * 100
+        f = numpy.float64(self.pl_operating_profit) / self.bs_net_assets * 100
         return f
 
     #　ＲＯＡを計算
     def roa(self):
-        f = self.pl_operating_profit / self.bs_total_assets() * 100
+        f = numpy.float64(self.pl_operating_profit) / self.bs_total_assets() * 100
         return f
